requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
gunicorn==21.2.0
python-dotenv==1.0.0
python-decouple==3.8